Fixed bug where control layer did not adher to simulated time ...
Note that this commit contains a bit of a hack. The call to
dls::ControlLayer::run sets the time library to use simulated time. This
is the wrong place to do it. The call Time::use_simulated_time(true)
should be moved to hyq_app_libraries.cpp. However, currently if it is
put there, the time is never updated. This is possibly due to something
with fastrtps, but still needs to be investigated before this hack can
be fixed.