test: added some non-automatic testing for fastrtps...
These tests are still exploratory, so it is not yet feasible to automate
them for future use.

More dummy controllers have been added to the framework. This makes it
easier to run multiple controllers at the same time and debug why they
are not all receiving communication from simulation

bare_dummy_controller.cpp is a simple test that runs a dummy controller
outside of the framework. This helps debug whether communication is
being lost due the framework or due to some other reason

blind_state_test is a simple blind state publisher. It helps determine
whether communication loss is due to the simulation publishing data, or
due to some other reason.

blind_state_sub is a simple blind state subscriber. It helps determine
whether communication loss is due to complexities arising in the
definition of controllers, or due to some other reason