WIP: new version for pub/sub...
Still attempting to fix the communication problem with fastrtps

Fastartps has changed their API. The publishers and subscribers in
namespace dls::version2 are using this new interface. If these
subscribers and publishers prove to work better, then the old ones will
slowly be removed and replaced with the new ones. When there are no more
of the old ones left, then these publishers and subscribers will be
taken out of the version2 namespace and put directly into the dls
namespace