WIP: fixing issue with multiple publishers to same topic...
Apparently, when using multiple publisher and subscribers in fastdds
through a single participant, fastdds will crash with a segfault if you
try and register the same topic twice. This is very annoying, because it
is not documented anywhere, and all of their examples are for a
participant per publisher, which is not how their library is supposed to
be used.

Currently, communication is not working at all, since all the publishers
have been ported to version2, but none of the subscribers have

A possible path forward would be to revert all the publishers back to
the previous version, then start porting the publisher and subscribers
in pairs incrementally to verify that everything keeps working as
intended