fix: dropped commands from console...
THIS IS A TEMPORARY FIX

There is an issue with the QoS settings of FastDDS whereby, when
registering multiple commands in quick succession, some of these are
dropped and not registered by the console. As a quick fix to enable the
rest of the development to proceed, a small time delay has been
introduced after the registration of a message. This should be removed
once the QoS settings have been stabilised.