WIP: gen dds message type for param server...
CMake is problematic. For some reason it was not possible to just use
`dls_add_message` inside the parameter server module. The reason for
this is not clear. In the top-level CMakeLists.txt, the main messages
for the framework are built with the call to:

	add_subdirectory(msg)

However, somehow this conflicts with calling the same function within
the parameter server module. Either of these can be done independently,
but when both of them are done together, cmake can't resolve the
dependencies.

Since the root cause cannot be resolved at this time, the parameter
server does not use the dls_add_message function, but instead implements
its own similar function.