removed the ability to load more than one simulator
It is not working and it may require too much time to debug. Now the
command "unloadSimulator dls_gazebo" works. Before you needed to launch
"unloadSimulator dls_gazebo_1" and each time you did "loadSimulator
dls_gazebo" a new instance of dls_gazebo was created in the console
function but another instance of the gazebo was not launched.