fixed issue with load/unloadController functions
After doing several time loadController trunk_controller and
unloadController trunk_controller, the loadController command
disappeared, as if the loading or unloading was not always working

The issue was caused probably by a deadlock condition due to a sleeping
thread holding a mutex. The levelWatcher thread was using
std::this_thread::sleep_for function (used in verifyLevel()) while
holding a mutex. This mutex is shared with loadController and
UnloadController DomainParticipant.