when tabbing, display commands only once they all match with commands_monitor of console layer
In order to display only the matched data readers (so only the commands ready to receive a message),
the list of commands needs to be populated only by the commands matching with commands_monitor.
Currently, in the FastDDS API it is not possible to check which data reader is matching with
a data writer from the data writer side.
It is not even possible to get the list of all data reader of domain/participant.
It is possible instead to get the current number of matched data reader.
So, to provide robustness about command execution, the list of commands is not empty only if all
the commands match with commands_monitor, otherwise an empty list is returned.
This is done by getting the commands names from the list of participants in the domain and
then checking if the number of participants is equal to the number of matched data reader.
In fact, it may happens that, even if the domain participant is running, one of its data reader
is not ready yet, leading to console command issue like "need to retype twice a command to execute it"