!!improved commands discovery
Each command is a Data Reader(DR) created from a domain participant
(DP). Only DP has a name. So to find a command, and so a command name,
which DR is belonging to which DP needs to be checked.
ConsolLayer::command_monitor data writer (DW) is publishing on the same
topic of DR, so:
-  when DW discover a DR, the InstanceHandle of DR is stored in a
vector
-  when DW discover a DP, the name-GuidPrefix tuple of the DP is
stored in a multi-map
- a command is found by matching discovered DR and DP GuidPrefix: if
they are the same, it means that DR belongs to DP, then the name is
stored