main.cpp.in REFACTORING
Supervisor is now a Layer, launched as an independent process
When launching the layers, you need now to explicitly launch the service
and the supervisor with: dls --super -lservice
Without --super, the dds servers are not launched, and the communication
cannot happen. Before launching the console layer, you need to have the dds server running!
It is a refactoring that precedes the startup procedure